# src/api/static/openapi.json
{
  "components": {
    "schemas": {
      "TransactionCreateSchema": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "amount",
          "description",
          "date",
          "category"
        ],
        "title": "TransactionCreateSchema",
        "type": "object"
      },
      "TransactionSchema": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "default": "2023-07-10T18:47:05.132157",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "default": "2023-07-10T18:47:05.132157",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "amount",
          "description",
          "date",
          "category"
        ],
        "title": "TransactionSchema",
        "type": "object"
      },
      "TransactionUpdateSchema": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          }
        },
        "title": "TransactionUpdateSchema",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Teste",
    "title": "Teste",
    "version": "1.0.0"
  },
  "openapi": "3.0.2",
  "paths": {
    "/transactions": {
      "get": {
        "description": "Retorna uma lista de transa\u00e7\u00f5es",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSchema"
                }
              }
            },
            "description": "Retorna o tenant criado"
          }
        },
        "summary": "Retorna uma lista de transa\u00e7\u00f5es"
      },
      "post": {
        "description": "Adiciona uma nova transa\u00e7\u00e3o",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCreateSchema"
                }
              }
            },
            "description": "Retorna a transa\u00e7\u00e3o criada"
          }
        },
        "summary": "Adiciona uma nova transa\u00e7\u00e3o"
      }
    },
    "/transactions/{id}": {
      "delete": {
        "description": "Remove uma transa\u00e7\u00e3o",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSchema"
                }
              }
            },
            "description": "Retorna o tenant criado"
          }
        },
        "summary": "Remove uma transa\u00e7\u00e3o"
      },
      "get": {
        "description": "Retorna uma transa\u00e7\u00e3o por id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionSchema"
                }
              }
            },
            "description": "Retorna o tenant criado"
          }
        },
        "summary": "Retorna uma transa\u00e7\u00e3o por id"
      }
    },
    "/transactions/{id}/tags": {
      "post": {
        "description": "Remove uma tag",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionCreateSchema"
                }
              }
            },
            "description": "Retorna a transa\u00e7\u00e3o editada"
          }
        },
        "summary": "Remove uma tag"
      }
    }
  }
}